Extract Provider
